[REFACTOR/#18] 패키지 리팩토링
