+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,13 @@
         android:theme="@style/Theme.TeumTeum"
         tools:targetApi="31">
         <activity
-            android:name=".SignUp.SignUpActivity"
+            android:name=".ui.singup.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".SignIn.LoginActivity"
+            android:name=".ui.signin.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".Splash.SplashActivity"
+            android:name=".ui.splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
