+++ app/src/main/java/com/example/teumteum/ui/singup/CompleteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.SignUp
+package com.example.teumteum.ui.singup
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.teumteum.R
 import com.example.teumteum.databinding.FragmentCompleteBinding
-import com.example.teumteum.databinding.FragmentOnBoardingNicknameBinding
 
 class CompleteFragment : Fragment() {
