+++ app/src/main/java/com/example/teumteum/ui/singup/OnBoardingNicknameFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.SignUp
+package com.example.teumteum.ui.singup
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.teumteum.R
-import com.example.teumteum.databinding.FragmentAgreementBinding
 import com.example.teumteum.databinding.FragmentOnBoardingNicknameBinding
 
 class OnBoardingNicknameFragment : Fragment() {
