+++ app/src/main/java/com/example/teumteum/ui/singup/SignUpActivity.kt
@@ -1,12 +1,9 @@
-package com.example.teumteum.SignUp
+package com.example.teumteum.ui.singup
 
 import android.os.Bundle
 import android.view.View
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.example.teumteum.R
 import com.example.teumteum.databinding.ActivitySignUpBinding
 
 class SignUpActivity : AppCompatActivity() {
