+++ app/src/main/res/layout/fragment_agreement.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignUp.AgreementFragment">
+    tools:context=".ui.singup.AgreementFragment">
 
     <!-- 제목 -->
     <TextView
