+++ app/src/main/res/layout/fragment_on_boarding_profile.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".SignUp.OnBoardingProfileFragment">
+    tools:context=".ui.singup.OnBoardingProfileFragment">
 
     <TextView
         android:id="@+id/title_tv"
